no more zombie processes; chms now align with naip tiles

+++ pointnet/hpo/rm_study.py
@@ -8,6 +8,10 @@ parser = argparse.ArgumentParser()
 parser.add_argument("--name", required=True)
 ARGS = parser.parse_args()
 
-
-shutil.rmtree(studypath(ARGS.name))
-shutil.rmtree(f"{ROOT}/pointnet/models/hpo/{ARGS.name}/")
+print(f"rm {ARGS.name}? y/[n]:", end=" ")
+if input() == "y":
+    shutil.rmtree(studypath(ARGS.name))
+    shutil.rmtree(f"{ROOT}/pointnet/models/hpo/{ARGS.name}/")
+    print("done")
+else:
+    print("cancelled")
